[WIP] V3 Advisory Format
This is a proposed new advisory format based on Markdown with embedded
TOML front matter.

Uses triple backquotes + `toml` for the front matter delimiter, which
renders correctly as syntax highlighted TOML on GitHub. This is a bit
unusual way to specify front matter (typically it's YAML with `---`
delimiters) but it seems to work ok.

The advisory `title` and `description` are moved out-of-band from the
rest of the advisory into the underlying Markdown document. This allows
it to render correctly on e.g. GitHub.