Make `Informational` a #[non_exhausive] enum.
Previously `Informational` was an enum exposed as a part of the public
API, which makes the addition of any new variants a SemVer-breaking
change.

This commit converts it to be #[non_exhaustive], allowing it to ensure
backwards compatibility even when new enum variants are added.

`warning::Kind` should get the same treatment, however
this commit leaves that for a follow-up.