Bump `cargo-lock` to v5.0; semver to v0.10; MSRV 1.41+
This commit goes back to re-exporting `Version` and `VersionReq` from
the `semver` crate, as opposed to using newtypes with a bunch of custom
code to implement prerelease handling.

This is likely a regression for RustSec/cargo-audit#17 and
RustSec/cargo-audit#30.

See also: steveklabnik/semver#172.

This needs to be addressed either by (re-)incorporating a similar fix,
or a fix upstream in the `semver` crate. It would be interesting to
re-examine the latter.