Refactor changes from `fetch` feature (#213)
The referenced PR added a `fetch` feature and made `Repository` into a
trait with `GitRepository` as the only implementer.

That's not really helpful for opening a local copy of the database
though, so this commit adds a proper `Database::open` method that can
load a database from a given directory.

It also removes the trait since it had no other impls and therefore
wasn't very helpful. We can perhaps circle back on renaming
`GitRepository` back to `Repository`, although it is more descriptive so
perhaps it should stay.