Add support for omitting leading `[advisory]` table
This change makes it possible to omit the leading `[advisory]` table
from the front matter.

It's a change it'd be nice to potentially land before freezing a stable
1.0 format.

While more discussion is needed, this change at least means if we do
want to go through with it, the next release of the `rustsec` crate will
support it.